+++ zone/client.cpp
@@ -1835,7 +1835,7 @@ bool Client::CheckIncreaseSkill(SkillUseTypes skillid, Mob *against_who, int cha
 	if (skillval < maxskill)
 	{
 		// the higher your current skill level, the harder it is
-		int16 Chance = 10 + chancemodi + ((252 - skillval) / 20);
+		int32 Chance = 10 + chancemodi + ((252 - skillval) / 20);
 
 		Chance = (Chance * RuleI(Character, SkillUpModifier) / 100);
 
@@ -1869,7 +1869,7 @@ void Client::CheckLanguageSkillIncrease(uint8 langid, uint8 TeacherSkill) {
 	int LangSkill = m_pp.languages[langid];		// get current language skill
 
 	if (LangSkill < 100) {	// if the language isn't already maxed
-		int16 Chance = 5 + ((TeacherSkill - LangSkill)/10);	// greater chance to learn if teacher's skill is much higher than yours
+		int32 Chance = 5 + ((TeacherSkill - LangSkill)/10);	// greater chance to learn if teacher's skill is much higher than yours
 		Chance = (Chance * RuleI(Character, SkillUpModifier)/100);
 
 		if(MakeRandomFloat(0,100) < Chance) {	// if they make the roll
@@ -3192,10 +3192,10 @@ uint16 Client::GetPrimarySkillValue()
 	return GetSkill(skill);
 }
 
-uint16 Client::GetTotalATK()
+uint32 Client::GetTotalATK()
 {
-	uint16 AttackRating = 0;
-	uint16 WornCap = itembonuses.ATK;
+	uint32 AttackRating = 0;
+	uint32 WornCap = itembonuses.ATK;
 
 	if(IsClient()) {
 		AttackRating = ((WornCap * 1.342) + (GetSkill(SkillOffense) * 1.345) + ((GetSTR() - 66) * 0.9) + (GetPrimarySkillValue() * 2.69));
@@ -3212,9 +3212,9 @@ uint16 Client::GetTotalATK()
 	return AttackRating;
 }
 
-uint16 Client::GetATKRating()
+uint32 Client::GetATKRating()
 {
-	uint16 AttackRating = 0;
+	uint32 AttackRating = 0;
 	if(IsClient()) {
 		AttackRating = (GetSkill(SkillOffense) * 1.345) + ((GetSTR() - 66) * 0.9) + (GetPrimarySkillValue() * 2.69);
